feat(harness): airtable.js shim + airtable-real.js rename
The real Airtable client moves byte-for-byte to airtable-real.js. The
new airtable.js is a 51-line shim that picks the real client or the
in-memory mock at module load time based on TEST_BACKEND. Production
sets no TEST_BACKEND, so the shim does a dynamic import of
airtable-real.js — same behaviour as before.

Two layers of footgun guard catch a stray TEST_BACKEND in Railway:
NODE_ENV=production + TEST_BACKEND set → process.exit(1) at boot, plus
a red-tinted log line whenever the mock is active so the mode is
visible in any deploy log.

The real code paths in airtable-real.js are unchanged. Every caller
imports from services/airtable.js and gets the same surface area
(list, getById, create, update, deleteRecord, atomicStockAdjust).
